requirements.txt
Django==4.1
Pyrebase4==4.8.0
Requests==2.32.3
